Feat: 핀 리스트와 핀 상세 페이지(더미) 연결
- 핀 상세 페이지 내 핀 이름, 메모 내용 등을 핀 데이터 변수로부터 받아 출력하도록 작성
- 날짜->Str 변환하는 DateFormatter를 헬퍼 파일로 별도 분리
