+++ ChagokChagok/ChagokChagok/ChagokChagokApp.swift
@@ -10,10 +10,13 @@ import SwiftUI
 @main
 struct ChagokChagokApp: App {
     let persistenceController = PersistenceController.shared
+    @StateObject private var dataController = DataController()
     
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            PinListView()
+                .environment(\.managedObjectContext, PersistenceController.preview.container.viewContext)
+                .environment(\.managedObjectContext, dataController.container.viewContext)
         }
     }
 }
