+++ ChagokChagok/ChagokChagok/Data/CoreData.xcdatamodeld/CoreData.xcdatamodel/contents
