+++ ChagokChagok/ChagokChagok/Data/DataController.swift
@@ -1,0 +1,21 @@
+//
+//  DataController.swift
+//  ChagokChagok
+//
+//  Created by Youngwoong Choi on 2022/06/11.
+//
+
+import Foundation
+import CoreData
+
+class DataController: ObservableObject {
+    let container = NSPersistentContainer(name: "CoreData")
+    
+    init() {
+        container.loadPersistentStores { description, error in
+            if let error = error {
+                print("Core Data failed to load: \(error.localizedDescription)")
+            }
+        }
+    }
+}
